Add inviscid module with inviscid-specific functions (#275)
